requirements.txt
# local package
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2
regex==2021.11.10
pandas
numpy
pandas-profiling==3.6.6
autoviz==0.1.604
scikit-learn
scipy==1.9.3
keras==2.13.1
torch==2.0.1
boto3==1.28.70
streamlit==1.27.1
fastapi==0.100.1
cookiecutter>= 1.4.0
langchain==0.0.300
openai==0.28.0
transformers @ git+https://github.com/huggingface/transformers@5936c8c57ccb2bda3b3f28856a7ef992c5c9f451
huggingface-hub==0.16.4
accelerate==0.23.0
bitsandbytes==0.41.1
ctransformers==0.2.27
gensim==3.6.0
einops==0.6.1



# external requirements
click
Sphinx
coverage
# awscli
# flake8
python-dotenv>=0.5.1
